+++ src/app/product/product.module.js
@@ -1,5 +1,5 @@
 (function () {
     'use strict';
 
-    angular.module('chpl.product', ['ui.bootstrap']);
+    angular.module('chpl.product', ['ui.bootstrap', 'chpl.common', 'chpl.loginServices']);
 })();